chore(app): Refactor history runs table stepper (#3992)
* move currentStep and steps into config

* refactor history stepper to accept other specs
